pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "devduck"
dynamic = ["version"]
description = "🦆 Self-adapting agent - one file."
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Cagatay Cali", email = "cagataycali@icloud.com"}
]
keywords = ["ai", "agent", "minimalist", "self-healing", "ollama", "strands-agents"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console", 
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11", 
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
requires-python = ">=3.10,<3.14"
dependencies = [
    "strands-agents",
    "prompt_toolkit",
    "strands-agents[ollama]",
    "strands-agents[anthropic]",
    "strands-agents[openai]",
    "strands-agents[otel]",
    "strands-agents-tools",
    "beautifulsoup4",
    "colorama",
    "websockets",
    "strands-mcp-server",
    "strands-google",
    "bedrock-agentcore-starter-toolkit",
    "bedrock-agentcore",
    "rumps; sys_platform == 'darwin'",
    "strands-mlx; sys_platform == 'darwin'",
    "eclipse-zenoh",
    "halo",
    "strands-agents[bidi-all]",
    "openai-whisper",
    "webrtcvad",
    "sounddevice",
]

[project.optional-dependencies]
rl = ["stable-baselines3[extra]", "gymnasium"]


[project.scripts]
devduck = "devduck:cli"
dev = "devduck:cli"

[project.urls]
Homepage = "https://github.com/cagataycali/devduck"
Repository = "https://github.com/cagataycali/devduck.git"
Documentation = "https://github.com/cagataycali/devduck#readme"
"Bug Tracker" = "https://github.com/cagataycali/devduck/issues"

[tool.setuptools]
packages = ["devduck", "devduck.tools"]

[tool.setuptools.package-data]
devduck = [
    "agentcore_handler.py", 
    "requirements.txt",
]

[tool.setuptools_scm]
write_to = "devduck/_version.py"
